third_party/WebKit/LayoutTests/platform/win/virtual/gpu/fast/canvas/canvas-currentColor-expected.txt

This is a testharness.js-based test.
FAIL Test that CanvasRenderingContext2D supports the 'currentColor' value. assert_equals: expected "255,0,0,255" but got "255,1,1,255"
Harness: the test ran to completion.
